+++ account_sdk/src/rpc_provider.rs
@@ -1,5 +1,1 @@
-use starknet::providers::JsonRpcClient;
 
-pub trait RpcClientProvider<T> {
-    fn get_client(&self) -> JsonRpcClient<T>;
-}
