+++ account_sdk/src/tests/one_file_all.rs
@@ -14,7 +14,7 @@ use starknet::{
 use crate::{
     deploy_contract::{get_account, CustomContract, CASM_STR, SIERRA_STR},
     deployer::{Declarable, TxConfig},
-    providers::{KatanaProvider, KatanaRunner, KatanaRunnerConfig, PREFUNDED},
+    providers::{KatanaProvider, KatanaRunner, KatanaRunnerConfig, FEE_TOKEN_ADDRESS, PREFUNDED},
     rpc_provider::RpcClientProvider,
     tests::find_free_port,
 };
